feat: add a status condition to ic for instrumentation details (#2408)
This adds a condition that describes why the workload is instrumented
(which kind of source), along with it's name for better troubleshooting
and traceability.

Examples:


![image](https://github.com/user-attachments/assets/16193fc3-3f95-4a5c-aa9b-1ab4572035b4)

The function in k8sutils now returns more info about the decision, which
can later be used for odigos describe and possibly be also persisted
with ic delete if we ever need finalizers for it.