delete ebpf instrumentation instances once odiglet shuts down (#2437)
Once the Odiglet is shutting down, call the `reporter.OnExit` callback
from the ebpf instrumentation manager.
When the odiglet is doing a rollout (in an upgrade for example) the ebpf
probes are unloaded and loaded - this creates a time period when there
is not active instrumentation - hence this change will remove the
instrumentation instances corresponding to these instrumentations in
that time period.