fix: destinations doc script - image sizes, and custom content demo text (#2443)
This pull request includes changes to various backend documentation
files, primarily focusing on updating image tags and adding detailed
instructions for creating accounts and obtaining access tokens. The most
important changes include updating the image tags to use a consistent
class and height, and adding detailed instructions for user actions.

Updates to image tags:

*
[`docs/backends-overview.mdx`](diffhunk://#diff-26115b197a8b9dd8ee351f05b2cade47da5acd788d74b9f962a325d3e1b919a2L15-R59):
Updated image tags for all backends to use a consistent class and
height.
*
[`docs/backends/appdynamics.mdx`](diffhunk://#diff-124c6f9b4d97583e112af5cf15db37c29ee22d49b620a9ea8baf810ad6dca8a2L10-R21):
Updated the image tag for AppDynamics to use a consistent class and
height.
*
[`docs/backends/awss3.mdx`](diffhunk://#diff-3a7ce85354b4b2eadec523fe8235cd31349580d1b3946e99bcf557b9d464abc6L10-R21):
Updated the image tag for AWS S3 to use a consistent class and height.
*
[`docs/backends/axiom.mdx`](diffhunk://#diff-183441861a9543c5648da503820d9b0c10a344e08138820fa4849f81a85beb00L10-R21):
Updated the image tag for Axiom to use a consistent class and height.
*
[`docs/backends/azureblob.mdx`](diffhunk://#diff-292b9c877fdd83df881a11992affe06aa78e14fbb1cac9449a266bdd4a0576e9L10-R21):
Updated the image tag for Azure Blob Storage to use a consistent class
and height.

Addition of detailed instructions:

*
[`docs/backends/appdynamics.mdx`](diffhunk://#diff-124c6f9b4d97583e112af5cf15db37c29ee22d49b620a9ea8baf810ad6dca8a2L10-R21):
Added detailed instructions for creating an account and obtaining an
access token.
*
[`docs/backends/awss3.mdx`](diffhunk://#diff-3a7ce85354b4b2eadec523fe8235cd31349580d1b3946e99bcf557b9d464abc6L10-R21):
Added detailed instructions for creating an account and obtaining an
access token.
*
[`docs/backends/axiom.mdx`](diffhunk://#diff-183441861a9543c5648da503820d9b0c10a344e08138820fa4849f81a85beb00L10-R21):
Added detailed instructions for creating an account and obtaining an
access token.
*
[`docs/backends/azureblob.mdx`](diffhunk://#diff-292b9c877fdd83df881a11992affe06aa78e14fbb1cac9449a266bdd4a0576e9L10-R21):
Added detailed instructions for creating an account and obtaining an
access token.
*
[`docs/backends/betterstack.mdx`](diffhunk://#diff-a37c129f12237e3b2bc3c543bc80c3440805bef299e2f27834df2e65935a0950L10-R21):
Added detailed instructions for creating an account and obtaining an
access token.