Use `registry.odigos.io` for images and update default `image-prefix` (#2438)
This updates the code to use `registry.odigos.io/*` instead of
dockerhub/`keyval/`. This also changes the default for the
`image-prefix` setting to `registry.odigos.io`

The reasoning for this is that previously, all of our images were on
dockerhub under the keyval org. For example: `keyval/odigos-odiglet`

This really evaluates to `docker.io/keyval/odigos-odiglet`, meaning that
even though the default value for `--image-prefix` is technically empty,
it is actually essentially defaulting to `--image-prefix=docker.io`.
This is why the new default `--image-prefix` is changing.

In addition, with the new registry, all of our images are hosted
directly under the main domain, such as
`registry.odigos.io/odigos-odiglet`

So, there is no need for us to add the `keyval/` prefix to our images
anymore. Similarly, there is no reason for users to add `keyval/` to
their own images when using their own registry. This is why we are
dropping it from the default image names.

These changes may be initially breaking to some users, but I think it is
a lot more intuitive:

* Default prefix: `docker.io`->`registry.odigos.io`
* Default image names `keyval/odigos-<component>` ->
`odigos-<component>`