cli: Support deleting/updating individual Sources by name (#2479)
Currently, the `odigos sources delete` and `odigos sources update`
commands only accept flags that identify Sources based on the Workload
they instrument. This is useful when the Source name might not be known
(such as Sources that were created with a random name by the UI), and it
also allows batch-deleting Sources (for example, deleting all Sources
that instrument a StatefulSet, etc)

This adds an optional argument to specify a Source explicity by its name
to these commands.

It also updates the help text (and docs) for these commands with
examples of usage in individual vs batch scenarios.