Remove "refetch" after CRUD operations in UI (#2512)
This PR solves issue https://github.com/odigos-io/odigos/issues/2511

---

Previously, after creating/updating/deleting via the UI - we would
perform a full refetch of rules (it does not have a watcher). In some
occasions, the data wouldn't be ready when performing the "refetch",
which made the UI seem like it wasn't synced with the latest - requiring
the user to refresh the page.

This PR solves this issue by returning the rule in the response of the
CRUD request. Thereby assuring the UI is updated instantly, and
preventing a follow-up request to fetch the data.

---

In addition:

- Added rules, actions, and destinations to the `usePaginatedStore`
store, preventing unnecessary queries when redirecting between pages.
- Removed "added" and "deleted" event batchers from destinations watcher
- not needed with new "data returned in CRUD request"