perf: separate instrumentation-instances from paginated-sources query (#2520)
Previously, when paginating sources in the UI, we would list the
`InstrumentationConfigs`, and for each config we would get the
`InstrumentationInstances` (just to add a "health| condition message to
the config).

Unfortunately, this approach costs a lot of time in some production
clusters. When fetching a batch of 10 apps, it would take less than 20ms
to list the configs, and it would take up to 20s to get the instances.

---

This PR improves the UI load-time, by separating
instrumentation-instances from the paginated-sources-query.

The new approach paginates sources much quicker, displaying them in the
UI - and allowing the user to work with them. Once all the sources are
there, we’ll fetch & update each source in the list with it's
instances-health condition message.

So the UI will have all sources - the sources will keep updating their
conditions, until we processed the last source in list, then we stop the
loop completely and listen to Server Sent Events (any modified event, we
refetch that source by ID - getting the updated data, including it’s
instances).