chore: remove unused image values from helm and add imagePrefix (#2531)
This PR is based on a feedback from user.

in helm `values.yaml` example file, we have image repository for each
component.
This is used for only 2 components to fill the `odigos-config` cm, but
are not actually consumed in the templates.
Also, since it only applies to few components but not all of them, it's
not actually useful for anything.

Also, the example files lacks the `imagePrefix` option.

While at it, I also removed the `deleteLangDetectionPods` value which is
unused.

Tested it locally by installing helm with:
- no value (manifests pull from `registry.odigos.io`)
- empty string (manifests pull from `registry.odigos.io`)
- invalid value - set to `foobar` (manifests contains
`foobar/odigos-...`) and fail to pull
- valid value - tested with `keyval` to point to the old dockerhub repo
(manifests contains `keyval/odigos-...` and works)