chore: remove deprecated "odigosconfiguration" CRD (#2534)
## Description

This CRD has been deprecated in #1405 . It's not being used since
`v1.0.93` and any version since then has already migrated it's content
to the new configmap alternative.

After enough time passed, it's no longer needed and can be cleaned up.

## How Has This Been Tested?

Tested the version without this CRD and verify it can run successfully.

I didn't test migration from versions `<v1.0.93` to a current version,
as this is not expected to work anyway.

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [x] Changes how Odigos interacts with Kubernetes
Remove a CRD which should be deleted by both helm and cli when upgrading
to new version

- [x] Modifies Odigos manifests (addressed in both CLI and Helm)
both CLI and helm will not install this CRD going forward, and will
delete it on upgrades.

- [x] Changes RBAC permissions
The only component that interacted with the CRD is the CLI, so no RBAC
permission changes are required.

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

~~- [ ] Users need to take action before upgrading~~
- [X] Automatic migration will modify existing objects (backward
compatible)
~~- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users
need to be aware of~~
- [X] Documentation updated accordingly - removed this CRD from the
odigos api part in the docs