GKE detect workaround (#2536)
Create a workaround to handle
https://github.com/GoogleCloudPlatform/opentelemetry-operations-go/issues/1026
figure out whether we're running on GKE at the startup of the autoscaler
(with a timeout of 2 seconds) - this should be removed once the issue
above is resolved and the collector dependency is updated.
In addition, the `resourcedetection` processor is updated to have a
timeout of 2 seconds.

## User Facing Changes

Not expected, users running on GKE should still see the resource
attributes.