feat: add internalTrafficPolicy data collection service for agents telemetry export (#2458)
setting `internalTrafficPolicy: Local` tells k8s to pass traffic to the
data collection pod on the same node.
ref:
https://kubernetes.io/docs/concepts/services-networking/service-traffic-policy/

### Motivation
Avoid using the `hostNetwork` on the data-collection - the
`internalTrafficPolicy` is the k8s recommended way of doing same node
traffic.
Using the host network can lead to issues with conflicting ports on the
host.

### Main changes
* Add a service for the data collection which is used to same node
traffic. It is used by the instrumentation agents and the odiglet to
pass the telemetry data.
* The `hostNetwork` is only added to the data collection if we decided
to not use the service (fallback to the approach used today).
* The new service will only be used on k8s version which are 1.26 or
later, since this is the version where this feature became GA (we can
decide to enable it from 1.22 which is the beta release of this version,
but this approach might cause issues if a cluster disabled that beta
feature).
* Add a `feature` package under `k8sutils` to allow us declare k8s
features we'd like to easily know if are enabled or not. Another example
where this is used is when setting the `maxSurge` field in a DaemonSet
RollingUpdate.
* fix: add `dnsPolicy: ClusterFirstWithHostNet` to odiglet in helm
charts - this is already present for the CLI, but was missing in the
charts. The reason it became important is since now the Odiglet needs to
be able to query DNS entries in addition to use the host network.

### Tests changes
* Add the ability to assert conditionally based on the k8s version. This
is done by passing the k8s minor version to chainsaw and using that
value to create conditions in the asserts.

### Future work

- [ ] enterprise PR to use the new service.
- [ ] Create a similar service for the odiglet - which will be used by
the OpAmp clients. - this will allow us to remove the `hostNetwork` for
the odiglet as well.
- [ ] `odigos diagnose` adjustments? not sure if needed.