test: assert less attributes with cli-upgrade before upgrading (#2541)
## Description

Attempt to fix the cli-upgrade test failure introduced with #2538 .

This PR makes it so that we don't assert instrumentation attributes when
installing odigos latest version as the first step for upgrading.

When we have breaking changes in agents or the attributes which we are
asserting, the test that uses the same assert file, will always break on
the latest published version.

I separated the files, so one yaml only checks if the applications are
instrumented, and the second one also assert the full attributes and
internal data which might break. cli-upgrade uses the first one for
"light" assertions only.

It will require that we apply any future updates to both files.

## How Has This Been Tested?

Attempting to make the cli-upgrade tests pass.

- [x] Updated e2e Tests

## Kubernetes Checklist

Only Tests

## User Facing Changes

Only Tests

---------

Co-authored-by: Ron Federman <73110295+RonFed@users.noreply.github.com>