Instances in UI - fetch all in a single query (#2537)
## Description

This PR changes the way we get the instances/healthy condition for
instrumented workloads.

Previously we would get them in the same query as paginated
instrumentation configs, then we changed it so we would 1st paginate the
instrumentation configs, and then get the instances for each workload
individually. This definitely improved load-time for instrumented
workloads in the UI, but still acted up with instances.

So now, we finish paginating instrumentation configs, and then we get
all instances in a single query.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

- [x] Introduces additional calls to the API Server (potential
performance impact)