operator+cli: Change how component image references are set (#2562)
## Description

This changes how component image references are defined. The goal of
this is for the operator to install components by their pinned SHA256
value (instead of the usual tag), which is a requirement for OpenShift
certification (to ensure operands refer to their actual certified image,
and it supports disconnected environments)

The main changes are in the cli. Instead of initializing each
component's image within each component's individual manager, it is done
once during the install command. The image references are persisted to
the odigos config, which is passed to each resource manager, where they
can consume the image names themselves.

This centralizes and simplifies the handling of image name cases
(default, enterprise, and ubi9).

This then allows the operator to set all of these image names by
environment variables (using the new config struct) which are generated
by `operator-sdk` to refer directly to the pinned digest for the
specified tags.

To handle this in each component, the `GetImageName` function is updated
to check if the image is a pinned SHA, and if so it does not append the
tag.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly