fix(instrumentor): fail to create source if webhook has error (#2592)
## Description

When creating new `Source` CRD to instrument a k8s workload, a mutating
webhook need to run in order to populate the object labels which are
required for instrumentor to work.

Up until now, the `FailurePolicy` was `Ignore` which means that the CR
would be created without the labels.

This caused the instrumentor to fail creating the
`InstrumentationConfig` and UI not to show the sources after they are
created.

This PR makes this case an error, so at least the user can retry the
operation (which might succeed if instrumentor was down temporarily).

## How Has This Been Tested?

- [x] Manual Testing

Scaled instrumentor to 0 replicas. created sources from UI and saw the
error. UI allows to add the sources again.
After starting instrumentor (with webhook) the UI is able to create new
sources successfully
Tested both with CLI and helm.

## User Facing Changes

Users that already have source objects without labels, would need to
delete those and create them again.
We might consider a followup PR to fix the issue automatically, but this
PR is simply not to have this happening for new workloads.