chore: delete crd dir before recreate (#2603)
Following #2534, when we run `make manifests` in `api` directory, it
creates manifests files based on the current go files, but any file that
is already there will not be deleted and left over in the directory.

This PR runs `rm -rf` on the crd directory before re-creating it's
content. This fix expose that `odigos.io_odigosconfigurations.yaml` is
not longer relevant. now to CI will verify this by running the make
target and checking git tree is clean

---------

Co-authored-by: Ben Elferink <ben.elferink@icloud.com>