feat: inject some distro env vars from manifest instead of code branching (#2598)
## Description

We want to retire some old code responsible for injecting some env vars,
to make the logic sit in the distro manifest instead of being set in
code.

This has the benefit of consolidating all the relevant per-distro
configurations into one file, and removing some of the tier-aware code
which should be migrated to enterprise instrumentor.

## How Has This Been Tested?

- [X] Updated e2e Tests - added all the env vars to the e2e instrumented
app assertions
- [X] Manual Testing - tested this in my local kind cluster. checked the
values in the manifest after rollout with new instrumentor, and made
sure all services are running, healthy, and emitting traces.

## Kubernetes Checklist

Just refactor of the code in the pods webhook. no changes in k8s api.

It's notable to mention that the new implementation checks for the
presence of an env var before setting it's value in the manifest.

## User Facing Changes

Internal refactors only. no user face change is expected.

---------

Co-authored-by: Ben Elferink <ben.elferink@icloud.com>