fix: propagate collectors backpressure to filelog receiver to slow it down  (#2698)
## Description

For telemetry arriving from agents when there is back-pressure, we never
want to propagate it back to the callers, so not to increase their
memory. but filelog receiver can pause and reduce the throughput on
pressure, and then resume when the pipeline is ready.

We never want to drop any logs on backpressure. 

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

It's kind of non-trivial to test, I tested it manually but we should
think how to improve testing of these kind of features.

## Kubernetes Checklist

No changes to k8s api.

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

No user facing changes. this fixes a bug in collectors stability and
dropping of data when it's unnecessary.