fix: migrate from honeycomb k8s-agent logs and metrics to otel (#2734)
## Description

https://github.com/honeycombio/honeycomb-kubernetes-agent is deprecated
and not recommended for collecting logs and metrics. this PR removes the
special case of honeycomb destination and migrate it to use the odigos
pipeline and telemetry collections just like any other destination.

## How Has This Been Tested?

- [x] Manual Testing

I configured an honeycomb destination and sent traces + logs + metrics
and confirmed the data is seen in honeycomb


## User Facing Changes

Existing users of honeycomb destination for logs and metrics will be
migrated to the new collection method which is aligned with the rest of
odigos