feat: add items per seconds for debug destination (#2736)
## Description

The debug destination currently logs all the telemetry items passing in
the pipeline.
For production environments with a lot of data, this can overwhelmed the
collector.

This PR adds an option to set lower volume for logging. for example one
item per second, protecting us from flooding the logs.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

No

## User Facing Changes

New feature in exporter allowing users to debug issues in pipeline