refactor: auto scaler controllers (#2735)
## Description

This PR aligns some aspects of the auto scaler controllers code to be
consistent with the rest of the repo, as well as be more readable and
organized:

- instead of having all controllers for both cluster and node collectors
in one directory, now each controller reside in the directory it is
relevant to.
- rename "datacollection" -> "node collector" and gateway to "cluster
collector"
- move manager and setup code to `manager.go` files to keep files more
focused and scoped, and have all initialization code in one place.

There are many more opportunities for improvements, but wanted to keep
this refactor scoped and not huge

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [x] Manual Testing

This is a refactored, so no changes in logic. I compiled and run it,
added and removed sources destinations and actions, and observed the
pipeline reacts

## Kubernetes Checklist

Split few controller runtimes controllers into 2 different objects. one
addressing node collector and one addressing cluster collector. This
should be the same as current logic.

## User Facing Changes

Internal refactor only. no changes for user.