fix: node collector resolve signals from cluster cg instead of destinations (#2737)
## Description

When auto scaler reconciled node collector configmap and daemonset, it
would use destinations for that.

Fixed:

- Destinations reconciler was not triggering the node reconciler, which
caused the cm not to update when a new destination is created that
should enable a new signal to be collected.
- exceptions for "span-to-metrics" were calculated from dest which might
not be in sync with actual pipeline (this was the case in victoria
metrics)
- remove unused parameter `dest` in sync daemonset function.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

Added another event filter for the collectors group reconciler which
should not have any meaningful overhead.

## User Facing Changes

Fix bug - if there is a tracing destination and one adds a logging
destination, it would apply the `filelog` receiver immediately. due to
current bug, users restart autoscaler as a workaround due to this bug