refactor: resolve service name in controller instead of webhook (#2739)
## Description

How we calculate the service name to use for each config?

- if the user set this in workload source object - use it
- otherwise - use the workload name as default ("coupon", "membership",
etc)

This PR moves this logic from the pods webhooik to the reconcilers,
insuring that the right value is always updated in the existing field
and is never empty.

Then we can set it only once, and don't need to resolve it downstream in
multiple consumers.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

Reduce work in webhook, and do it only few times in controller instead
of again and again for each pod.
This removes a `Get` request to the cache during webhook which is a good
improvement 🎉

## User Facing Changes

Internal refactor