fix: resolve also workload names in k8s attributes processor (#2742)
## Description

The k8sattributes processor is designed for our spans, where spans
generated from odigos always contains `k8s.deployment.name`,
`k8s.daemoneset.name` or `k8s.statefulset.name`.

However, for logs collected with `filelog` receiver, those attributes
are missing and are not enhanced by the processor.
This PR checks if one of the k8sattributes actions is configured to
collect logs, and add those attributes by default if they are.

## How Has This Been Tested?

- [x] Manual Testing

## Kubernetes Checklist

This change requires to resolve deployments etc in node collector. we
need to do it anyway as logs are not very useful without this info of
what workload originated it. the `k8sattributes` processor uses informer
and caches data internally and is a contrib processor which we already
use and recommend.

## User Facing Changes

Uses which collect logs and have the `k8sattributes` action enabled on
logs will see one new attribute for the workload name.