fix: mount /mnt/var/log into node collector (#2750)
## Description

Odigos reads logs from k8s node fs in `/var/log`. in some systems, this
path is a symlink to `/mnt/var/log`, so we need to mount this as well
for the logs collection to work.

This will not cover all possible cases, but the hope is that
`/mnt/var/log` is common enough to cover most non-trivial cases.