Handle empty JAVA env vars as if they are not defined (#2773)
Following https://github.com/odigos-io/odigos/pull/2667, we append our
value only to one env var if we detect a user defined value. That PR
also covers the case where the env var (for example JAVA_OPTS) is
declared by the user and have an empty value. For that case, we'll add
our value only to the JAVA_OPTS. Most (if not all) the JDKs don't looks
for the JAVA_OPTS env var for agent injecting - hence this results in
the agent not being injected.
We should completely remove JAVA_OPTS in a future PR, this is done as a
quick fix.