ci: notify slack on success / failure of offset check pre release (#2777)
When we push a new tag for release, and the offset PR in enterprise repo
is not merge, the release stops but there isn't any message in slack for
that.

This PR adds a success / error messages so we can be aware + a link to
the offsets PR for quick reaction with minimal toll.

It's mostly chatGPT and not trivial to test, so we need to keep an eye
in the next few releases.

We also need to think if it's easy to re-trigger the release in OSS
after the offsets PR is merged (in case one stopped due to missing
offsets) but that's a task for future work.