fix: Align helm to CLI by setting default imagePrefix to our registry (#2782)
This PR fix a bug in which we didn't set the `imagePrefix` config option
when installing with helm.
This cause weird behaviors for example:
When installing via helm and then running `odigos config set ...` the
config command would read the current config - see an empty image prefix
and would return an image like "odigos-ui/v1.0.177" - which is invalid
and will result in `ImagePullBackOff`.

Adding the util imagePrefix helm function, allowing us to use it in:
1. When setting the `odigos-config` values.
2. When calling the util `imageName` which is now using the
`imagePrefix` util.

Writing the imagePrefix correctly to the config map - allows
compatibility with the CLI commands.