Support setting a NodeSelector on Odigos components (#2764)
## Description

This adds the option to the Operator to support installing Odigos only
on specific nodes. This can be extended to CLI and Helm too, if desired.

This may seem odd, because it means that Odigos will only be able to
instrument applications that are scheduled onto the same node(s) that
Odigos is installed on. Because of that, this likely won't get used
often.

But we have seen a user that is running in a shared, but restricted,
environment where they have control over their own applications, but not
the underlying node infrastructure. They also have a very restricted
quota in their namespace.

In this case, the Odiglet Daemonset is consuming their quota because it
is attempting to schedule an odiglet on every (shared) node, even though
they only need one. In this case, it is up to the user to ensure their
application is also scheduled on the proper node, but this gives them
the option to limit how many Odiglets are scheduled.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly