openshift: Check if SEManage rule already exists (#2805)
## Description

SELinux doesn't seem to consistently handle re-applying existing rules
with `semanage`. It can sometimes result in an error that currently
prevents the Odiglet from starting up. This PR does 2 things for RHEL
hosts:

1. First check if an selinux rule exists for Odigos. If it does, don't
both trying to re-apply
2. If we do hit some (unknown) error anyway, log the error and don't
crash the Odiglet. This means we can still start up as a best effort in
error cases, with the tradeoff that Odigos may not work if there was a
different issue.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [x] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly