fix: remove sources in uninstall and wait for the pods to roll out cleanly (#2784)
## Description

Fixing a bug in which the instrumented pods were not rolled out during
uninstall to remove the instrumentation.
There was an existing code that removes the Sources, but that was done
after the instrumentor deployment was removed - hence, the controller
was missing and the pods remained "instrumented".

This PR adds the sources deletion as the first step of the uninstall
process, and potentially waits for all the pods to roll out before
continuing to delete the rest of the Odigos resources.

Waiting for pods to roll out without odigos modifications can take time
in larger clusters, hence a `--no-wait` flag is added which skips this
waiting phase.

On the other hand, when deleting the Sources we have to wait for the
`instrumentor` to remove them. Not doing that can lead to garbage
sources being left behind, and even the Source CRD can remain after an
uninstall (since once a CRD is deleted, a finalizer is added to only
remove it once all its instances are deleted).

Co-authored-by: Amir Blum <amirgiraffe@gmail.com>