feat: add configuration option for logs symlink target directory (#2822)
## Description

Odigos will attempt to read logs from `/var/log`

Sometimes, this is a symlink to some other directory like
`/mnt/var/log`. In this case, odigos will not be able to follow this
link in the node collector container since it's not mounted.

This PR adds option to configure the target for the `/var/log` symlink
so odigos is able to mount it into the collector pod as well for logs
collection to work.

## How Has This Been Tested?

<!-- Describe the tests you ran and how you verified your changes. -->

- [ ] Added Unit Tests
- [ ] Updated e2e Tests
- [ ] Manual Testing
- [ ] Manual Load Test

## Kubernetes Checklist

<!-- If this PR affects how Odigos interacts with Kubernetes, check the
relevant boxes below and provide more details -->

- [ ] Changes how Odigos interacts with Kubernetes
- [ ] Introduces additional calls to the API Server (potential
performance impact)
- [ ] New Query/feature supported in all the k8s versions supported by
Odigos
- [ ] Modifies Odigos manifests (addressed in both CLI and Helm)
- [ ] Changes RBAC permissions

## User Facing Changes

<!-- Any changes that users will notice or need to be aware of -->

- [ ] Users need to take action before upgrading
- [ ] Automatic migration will modify existing objects (backward
compatible)
- [ ] Changes UI, CLI, or K8s Manifests aspects in a way that users need
to be aware of
- [ ] Documentation updated accordingly